refactor(bridge): host normalize_auth_header in one place
`normalize_auth_header` — which turns a configured `ws_auth_header`
into the verbatim `Authorization` value — was duplicated in
`config::compile` and `core::acceptor`, each with a comment noting the
two must stay in sync.

Move it to `siphon-ai-bridge`, the crate that owns the wire-header
behavior (`conn::build_request` emits the header). `core` already
depends on `bridge`; `config` gains a direct dependency. The unit
tests move with the function. One implementation, no sync-by-hand.
