+++ crates/core/src/acceptor.rs
@@ -63,8 +63,8 @@ use sip_uac::integrated::IntegratedUAC;
 use sip_uas::integrated::{AcceptInviteAsyncOutcome, IntegratedUAS};
 use sip_uas::{NegotiatedSessionTimer, SessionTimerPolicy, UserAgentServer};
 use siphon_ai_bridge::{
-    AudioEncoding, AudioFormat, BridgeConfig, CallId as BridgeCallId, Direction, DisconnectReason,
-    OutgoingEvent, SipMeta, StartMsg, PROTOCOL_VERSION,
+    normalize_auth_header, AudioEncoding, AudioFormat, BridgeConfig, CallId as BridgeCallId,
+    Direction, DisconnectReason, OutgoingEvent, SipMeta, StartMsg, PROTOCOL_VERSION,
 };
 use siphon_ai_cdr::{
     AudioInfo as CdrAudioInfo, CdrRecord, CdrSinkHandle, Direction as CdrDirection, NullSink,
@@ -341,33 +341,6 @@ pub fn build_bridge_config(
         auth_header,
         connect_timeout,
     })
-}
-
-/// Normalize a configured WS auth header into the full
-/// `Authorization` value the bridge emits verbatim.
-///
-/// - `"Bearer xxx"`, `"Basic abc"`, `"Digest …"` → returned as-is
-///   (any RFC 9110 scheme works; the WS conn layer sends what we
-///   hand it without reformatting).
-/// - `"xxx"` (a bare token with no whitespace) → `"Bearer xxx"`.
-///   Preserves the historic UX where operators wrote just the
-///   token for Bearer-auth WS servers.
-///
-/// Previously this function was named `strip_bearer_prefix` and
-/// returned only the token portion of "Bearer xxx", leaving the
-/// caller (`bridge/conn.rs`) to re-prepend "Bearer " unconditionally
-/// when emitting the header. That broke any non-Bearer scheme:
-/// `"Basic abc"` would survive the strip unchanged, then become
-/// `"Bearer Basic abc"` on the wire.
-fn normalize_auth_header(header: &str) -> String {
-    let trimmed = header.trim();
-    // A bare token has no inner whitespace; an "Authorization"
-    // header always has a scheme keyword followed by a space.
-    if trimmed.contains(char::is_whitespace) {
-        trimmed.to_string()
-    } else {
-        format!("Bearer {trimmed}")
-    }
 }
 
 /// Resolve the codec list for a matched route. The route's
@@ -2640,41 +2613,6 @@ a=sendrecv\r\n",
         assert_eq!(outcome.answer_direction, None);
         assert_eq!(outcome.answer_sdp, cached_answer_pcmu());
         assert_eq!(outcome.remote_addr, None);
-    }
-
-    // ─── normalize_auth_header ─────────────────────────────────────
-
-    #[test]
-    fn normalize_auth_header_passes_bearer_through() {
-        assert_eq!(normalize_auth_header("Bearer abc"), "Bearer abc");
-    }
-
-    #[test]
-    fn normalize_auth_header_passes_basic_through() {
-        assert_eq!(
-            normalize_auth_header("Basic dXNlcjpwYXNz"),
-            "Basic dXNlcjpwYXNz"
-        );
-    }
-
-    #[test]
-    fn normalize_auth_header_passes_digest_through() {
-        assert_eq!(
-            normalize_auth_header("Digest username=\"foo\""),
-            "Digest username=\"foo\""
-        );
-    }
-
-    #[test]
-    fn normalize_auth_header_promotes_bare_token_to_bearer() {
-        // Historic UX: operators wrote just the token for Bearer auth.
-        assert_eq!(normalize_auth_header("xyz123"), "Bearer xyz123");
-    }
-
-    #[test]
-    fn normalize_auth_header_trims_outer_whitespace() {
-        assert_eq!(normalize_auth_header("  Bearer abc  "), "Bearer abc");
-        assert_eq!(normalize_auth_header("  xyz123  "), "Bearer xyz123");
     }
 
     // ─── sdp_audio_remote_addr ─────────────────────────────────────
